upd: library models – Librarian
The idea is to have a user with unreachable for normal users id, which is vk_id=0 to make a public bookshelf. The "User" with vk_id=0 is a librarian. The books from his bookshelf can be added to LibrarianJournal to indicate, that somebody copied them in their bookshelf.
There is can be only one public bookshelf in the system.
There is can be only one Librarian in the system as well.
The amount of public books in system is unlimited.